https://github.com/ldleman/yana-server/issues/5 Empeche l'utilisateur d'effacer le dernier utilisateur ce qui est provoque des erreurs qui bloque yana-server Bonus point: les notifications se ferme avec un fade-out

+++ action.php
@@ -66,10 +66,17 @@ switch ($_['action']){
 	case 'delete_user':
 		if(!$myUser->can('user','d')) exit('ERREUR: Permissions insuffisantes.');
 		$userManager = new User();
-		if(isset($_['id'])){
+		$NbUsers = $userManager->countUsers();
+
+		
+		if(isset($_['id']) && $NbUsers > 1){
 			$userManager->delete(array('id'=>$_['id']));
+			header('location:setting.php?section=user');
 		}
-		header('location:setting.php?section=user');
+		else
+		{
+		header('location:setting.php?section=user&error=Impossible de supprimer le dernier utilisateur.');
+		}
 	break;
